Rename SelectFiat->EditName
